softcode manager in bpod objects + tests
